% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion1.R
\name{saludo}
\alias{saludo}
\title{Funcion Saludo}
\usage{
saludo(n)
}
\arguments{
\item{n}{nombre del usuario.}
}
\value{
saludo al usuario.
}
\description{
Damos un saludo matutino al usuario
}
\examples{
saludo("Luisa")
saludo("Mercedes")
}
